requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/dev.txt requirements/ci.txt requirements/dev.in
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.3            # via black
attrs==19.3.0             # via black
babel==2.8.0              # via sphinx
beautifulsoup4==4.8.2
black==19.10b0
bumpversion==0.5.3
certifi==2019.11.28
chardet==3.0.4
click==7.0                # via black, pip-tools
commonmark==0.9.1         # via recommonmark
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
django-appconf==1.0.3
django-axes==5.2.2
django-choices==1.7.1
django-cors-middleware==1.5.0
django-debug-toolbar==2.1
django-extensions==2.2.6
django-filter==2.2.0
django-ipware==2.1.0
django-redis==4.11.0
django-relativedelta==1.1.0
django-solo==1.1.3
django-webtest==1.9.7
django==2.2.9
djangorestframework-camel-case==1.1.2
djangorestframework==3.9.4
docutils==0.16            # via recommonmark, sphinx
drf-nested-routers==0.91
drf-yasg==1.16.0
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==4.0.0
flake8==3.7.9
freezegun==0.3.14
gemma-zds-client==0.13.1
idna==2.8
imagesize==1.2.0          # via sphinx
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
jinja2==2.11.0
markdown==3.1.1
markupsafe==1.1.1
mccabe==0.6.1             # via flake8
oyaml==0.9
packaging==20.1           # via sphinx
pathspec==0.7.0           # via black
pip-tools==4.4.0
psycopg2==2.8.4
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygments==2.5.2           # via sphinx
pyjwt==1.7.1
pyparsing==2.4.6          # via packaging
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.10.5
pytz==2019.3
pyyaml==5.3
raven==6.10.0
recommonmark==0.6.0
redis==3.3.11
regex==2020.1.8           # via black
requests==2.22.0
ruamel.yaml==0.16.6
six==1.14.0
snowballstemmer==2.0.0    # via sphinx
soupsieve==1.9.5
sphinx-rtd-theme==0.4.3
sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
sqlparse==0.3.0
tablib[xls]==1.0.0
tblib==1.6.0
text-unidecode==1.3
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
uwsgi==2.0.18
vng-api-common==1.0.36
waitress==1.4.2
webob==1.8.6
webtest==2.0.33
xlrd==1.2.0
xlwt==1.3.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
